WIP sur GLSL
Grosses modif pour migrer les shaders des shadow volumes et le solide
photométrique. J'ai eu un gros souci à cause de la locale française qui me
faisait lire des ',' à la place des '.'. J'ai résolu temporairement le
problème en faisant un setlocale(LC_NUMERIC,"C") dans l'appli. J'ai essayé
de le faire temporairement mais cela ne marchait pas. Par contre du coup,
il a fallu que je modifie également les fichiers de config pour remplacer
les ',' par des '.'.
Les shadow volumes ne marchent pas encore, le solide photométrique seul non
plus, mais je préfère commiter maintenant car les changements sont déjà
conséquents. J'ai en outre corrigé quelques bugs d'affichage, notamment
sans le DP et le glow.

+++ src/common.hpp
@@ -105,8 +105,8 @@ public:
   SolverConfig globalField;
   uint nbFlames;
   FlameConfig *flames;
-  double fatness[4];
-  double extrudeDist[4];
+  GLfloat fatness[4];
+  GLfloat extrudeDist[4];
   double gammaCorrection;
 };
